pyproject.toml
[tool.poetry]
name = "python"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^9.4.0"
pytesseract = "^0.3.10"
opencv-python = "^4.7.0.72"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
